+++ app/Filament/Resources/SrsCardResource/Pages/CreateSrsCard.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Filament\Resources\SrsCardResource\Pages;
+
+use App\Filament\Resources\SrsCardResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateSrsCard extends CreateRecord
+{
+    protected static string $resource = SrsCardResource::class;
+}